backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
python-dotenv==0.19.0
google-generativeai==0.5.2
openai==1.3.5
Pillow
moviepy==1.0.3
requests==2.28.1
python-multipart==0.0.5
Werkzeug==2.0.1
redis==5.0.8
mysql-connector-python==9.0.0
SQLAlchemy==2.0.25
